feat[courses-service]: add full base model exclude exercises

+++ courses-service/src/main/java/ru/just/courses/service/ThemeService.java
@@ -5,15 +5,18 @@ import lombok.RequiredArgsConstructor;
 import org.hibernate.Session;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.just.courses.controller.exception.MethodNotAllowedException;
 import ru.just.courses.dto.CreateThemeDto;
 import ru.just.courses.dto.ThemeDto;
+import ru.just.courses.model.Module;
 import ru.just.courses.model.theme.Theme;
 import ru.just.courses.model.theme.content.TextThemeContent;
 import ru.just.courses.repository.ModuleRepository;
 import ru.just.courses.repository.TextThemeContentRepository;
 import ru.just.courses.repository.ThemeRepository;
-import ru.just.courses.repository.exceptions.EntityNotFoundException;
+import ru.just.courses.repository.exception.EntityNotFoundException;
 import ru.just.courses.repository.projection.TextThemeContentProjection;
+import ru.just.securitylib.service.ThreadLocalTokenService;
 
 import java.io.*;
 import java.sql.Clob;
@@ -27,6 +30,7 @@ public class ThemeService {
     private final ThemeRepository themeRepository;
     private final ModuleRepository moduleRepository;
     private final TextThemeContentRepository textContentRepository;
+    private final ThreadLocalTokenService tokenService;
     private final EntityManager entityManager;
 
     public Optional<ThemeDto> findThemeById(Long themeId) {
@@ -34,22 +38,27 @@ public class ThemeService {
     }
 
     public ThemeDto createTheme(CreateThemeDto dto) {
-        if (!moduleRepository.existsById(dto.getModuleId())) {
+        final Optional<Module> moduleOptional = moduleRepository.findById(dto.getModuleId());
+        if (moduleOptional.isEmpty()) {
             throw new NoSuchElementException("module with specified id doesn't exists");
         }
+        checkCourseAuthor(dto.toEntity().withModule(moduleOptional.get()));
+
         final Theme theme = themeRepository.save(dto.toEntity());
         return new ThemeDto().fromEntity(theme);
     }
 
     public void deleteById(Long themeId) {
+        checkCourseAuthor(themeRepository.findById(themeId).orElseThrow(() -> new EntityNotFoundException("Theme with specified id not found")));
         themeRepository.deleteById(themeId);
     }
 
-    public void addTextContentToTheme(Long themeId, InputStream inputStream, Integer textLength) {
+    public void addTextContentToTheme(Long themeId, InputStream inputStream, long length, Integer ordinalNumber) {
+        checkCourseAuthor(themeRepository.findById(themeId).orElseThrow(() -> new EntityNotFoundException("Theme with specified id not found")));
         Session session = entityManager.unwrap(Session.class);
         BufferedReader br = new BufferedReader(new InputStreamReader(inputStream));
-        Clob text = session.getLobHelper().createClob(br, textLength);
-        textContentRepository.save(new TextThemeContent(text, themeId));
+        Clob text = session.getLobHelper().createClob(br, length);
+        textContentRepository.save(new TextThemeContent(text, themeId, ordinalNumber));
     }
 
     @Transactional
@@ -61,7 +70,7 @@ public class ThemeService {
 
     private void writeResponse(TextThemeContentProjection themeContent, OutputStream out) {
         byte[] buffer = new byte[1024];
-        try (BufferedInputStream br = new BufferedInputStream(themeContent.getText().getAsciiStream())){
+        try (BufferedInputStream br = new BufferedInputStream(themeContent.getText().getAsciiStream())) {
             int realSize;
             while ((realSize = br.read(buffer)) != -1) {
                 out.write(buffer, 0, realSize);
@@ -69,6 +78,12 @@ public class ThemeService {
             out.flush();
         } catch (IOException | SQLException e) {
             throw new EntityNotFoundException("Не удалось передать текстовый контент");
+        }
+    }
+
+    private void checkCourseAuthor(Theme theme) {
+        if (!theme.getModule().getCourse().getAuthorId().equals(tokenService.getUserId())) {
+            throw new MethodNotAllowedException("You are not a course author");
         }
     }
 }
